Modify: Split rich text and i18n part
